Implement JSON Web Token based authentication
Typically, web servers use session-based stores to keep track of
authenticated clients. However, that method doesn't scale, since every
request would have to be routed to a server instance that keeps track of
the session that the request belongs to.

Our implementation uses JWTs and is therefore stateless. Upon successful
account creation or login, the server generates a JWT and returns it to
the client. The client is responsible for storing the token, and
including it in subsequent requests for authorization.

All API endpoints other than POST /login and POST /users requires token
authentication.

Reference: https://github.com/themikenicholson/passport-jwt

[Finishes #122731901] [#122423511] [Finishes #123203941]