Fix broken error handling logic
Express requires a function signature of 'app.use(function(err, req, res,
next))' for error handlers. The previous implementation of the handlers
didn't work because the 'next' argument was omitted to satisfy the
jshint checks.

We will also be setting err.status for all errors where we desire to set
a custom response code. We'll address the other changes in future
commits. [#126363131]