Create Trip model and add first POST and GET routes
The trip is the new main entity of the Journey app. Users can create
trips and add places they want to go to these trips. When they're in
their trips, they can use the app to find their way between different
places, discover interesting new places to visit, and so forth.

This commit introduces the Trip model to the backend and implements
functionality for creating a new trip, as well as for fetching all the
trips that a user has created. [#127139351]