Implement Trip Idea model and initial routes
A trip contains a list of ideas. Each idea contains information about a
place the user wants to visit. Most of the fields in the schema will
come from information from the Google Place API, which the clients are
responsible for calling.

This first commit supports API endpoints for adding an idea to a trip,
as well as getting a trip's list of ideas.

[#127246889]