Change trip destination schema and add API routes
Instead of storing trip destinations as a list of strings, we devise a
custom schema for each destination, allowing us to store a Google Place
ID, GPS coordinates, and the type of place. One example use case for
this data is to identify the bounding box to display a map of the trip
destinations for the user. [Finishes #127398901]