Implement API for creating trip entries
Allows trip entries to be inserted into trip days. Must specify a trip
ID and a trip day ID.

Also added logic to add/remove days from the trip plan if the user
changes the duration of a trip. Any entries that are in days to be
deleted are moved to the latest entry that is not deleted.

[#Finishes #127872563]
[#127443661]