Remove uniqueness restriction on trip ideas
An earlier commit included the logic that prevented two or more trip
ideas from referencing the same Google Place ID. The intention was to
help the user keep a succint list of ideas without duplication. However,
this restriction increases the complexity of the trip ideas model. If
un-planning a trip entry re-adds it as a trip idea, then the user could
potentially lose data. Also, this earlier restriction assumes that users
want only one idea for each place they want to visit. It may not be
true.

This commit undoes the logic, and helps with the implementation of trip
entry un-planning in the future.

[#127443743]