Support adding a trip entry based on a trip idea
If the trip entry creation API is called with a valid idea in the
request body, the backend creates the entry using the contents of the
idea and removes the idea from the list.

This change also adds support for adding an entry into a specific
position in the entries list for the specified day.

[#127443743]