Save username to new user account if provided
A previous commit no longer captured the username provided in the HTTP
request for creating a new account. This commit adds it back, and
updates the user schema to ensure that the username field is not
required. An existing pre-save check already ensures that all usernames
saved to the collection are unique. [#129300501]